backend/solana-backend/requirements.txt
anyio==3.6.1
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
construct==2.10.67
construct-typing==0.5.2
fastapi==0.85.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
solana==0.27.0
solders==0.9.0
starlette==0.20.4
types-cachetools==4.2.10
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
websockets==10.3
